backend/requirements.txt
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23
alembic==1.13.0
Pillow==10.1.0
python-dotenv==1.0.0